Fix ddim for Mac: #264

+++ comfy/ldm/models/diffusion/ddim.py
@@ -18,7 +18,7 @@ class DDIMSampler(object):
     def register_buffer(self, name, attr):
         if type(attr) == torch.Tensor:
             if attr.device != self.device:
-                attr = attr.to(self.device)
+                attr = attr.float().to(self.device)
         setattr(self, name, attr)
 
     def make_schedule(self, ddim_num_steps, ddim_discretize="uniform", ddim_eta=0., verbose=True):
